Added DEBUG logging levels: DEBUG_BASIC, DEBUG_DETAILED, DEBUG_VERBOSE
create_new_slideshow.py
- removed square grouping so that square images get grouped according to square_img config setting (Landscape or Portrait)
- added timestamp support for slideshow creation

import_photos.py
- changed await call to new process_image() method

mat_image.py:
- added config value for resampling method (BICUBIC, LANCZOS, ...)

model.py:
- new debug levels
- new default config

process_images.py:
- made smart cropping configurable, changed it keep full image and use letterboxing for now
- added config value for resampling method (BICUBIC, LANCZOS, ...)
- rewrote process_image to always return pyvips image object instead of file
  (previously it switched from passing file to image and back, making it error-prone)
  added checks to make sure file didn't get deleted until processing complete
  removed now redundant get_exif_corrected_dimensions()
- disabled separate process_single_image_async() and made process_image() async
- cleaned-up and stream-lined the process

viewer_display.py:
- changed __matter.mat_image(im) to __matter.mat_image([im]) so it passes a list, which is what it expects
  which made it possible to enable matting and disable always filling to bounds (fixed texture creation error)

stop.sh:
- fixed error when unbuffer process not found

deleted incorrect .flake8 configuration file, modified tox.ini with [flake8] section to eliminate
E203 (whitespace error that conflicts with Black)
C901 (too complex)

+++ src/picframe/scripts/stop.sh
@@ -23,7 +23,7 @@ if [ -n "$python_process_id" ]; then
     fi
 
     # Check if still running
-    if ps -p $unbuffer_process_id > /dev/null; then
+    if [ -n "$unbuffer_process_id" ] && ps -p $unbuffer_process_id > /dev/null; then
         echo "Unbuffer process did not terminate gracefully, killing forcefully..."
         sudo kill -9 $unbuffer_process_id
     else
